refactor(skill): split skill.rs into module, add phase 2 stubs
Pre-work for phase 2 basic skills (Translate / Polish / Summarize /
Compose). Splits the 563-line monolithic skill.rs into per-skill
files so multiple agents can work on different skills in parallel
without merge conflicts on a single file.

skill/mod.rs
  - Skill trait + ExecutionTarget / Privacy / SkillOutput
  - SkillRegistry with register / get / dispatch / tool_definitions
  - slugify() helper (memory-skill shared util)
  - Module declarations + pub use re-exports

skill/echo.rs        EchoSkill (sanity check)
skill/remember.rs    RememberSkill (existing)
skill/recall.rs      RecallMemorySkill (existing)
skill/forget.rs      ForgetMemorySkill (existing)
skill/edit.rs        EditMemorySkill (existing)

skill/translate.rs   TranslateSkill stub — execute() returns "not impl"
skill/polish.rs      PolishSkill stub — same
skill/summarize.rs   SummarizeSkill stub — same
skill/compose.rs     ComposeSkill stub — same

Each phase 2 skill takes Arc<dyn LlmProvider> in constructor (they
all need to make their own chat call to do the actual work — translate,
polish, etc.). This is parallel to memory skills taking Arc<dyn MemoryStore>.

Public re-exports kept identical to before — mori-tauri's main.rs
imports are unchanged.

cargo check --workspace ✓
cargo test -p mori-core --lib ✓ (existing 2 memory tests still pass)

Next: 4 parallel agents fill in each phase 2 skill.
